backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
openai==1.3.0
requests==2.26.0
redis==4.0.2
pydantic==1.8.2
python-jose==3.3.0
aiohttp==3.8.1
fastapi-cache2==0.1.8
prometheus-client==0.11.0 